+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Recommendation: for ease of reading, use the following order:
 ## Unreleased
 ### Changed
 - Updated `sqlx` crate to `0.8.6`, Vol. 2
+### Fixed
+- `web3-wallet` authorization provider: interactive login use case support (Device Flow) 
 
 ## [0.240.0] - 2025-06-02
 ### Added